Add render pass logic to sample multiple textures
This change adds a new struct to sample a number of rendered frame buffers with textures. It's independent of `Canvas2D` and just requires to render the scene to a texture. Thereafter the textures are cycled in a `VecDeque` (double ended queue), keeping a number of rendered frames in the list, while the current frame is written. This way multiple images can be sampled in the fragment shader that combines pixel data from these textures.